+++ src/ITempFileSystemOptions.ts
@@ -1,0 +1,17 @@
+import { ITempNameOptions } from "./ITempNameOptions";
+
+/**
+ * Provides options for creating a new temporary file-system entry.
+ */
+export interface ITempFileSystemOptions extends ITempNameOptions
+{
+    /**
+     * A value indicating whether the file-entry should be kept after exiting the process.
+     */
+    Keep?: boolean;
+
+    /**
+     * The mode to create the file-system entry with.
+     */
+    Mode?: number;
+}