+++ src/TempResult.ts
@@ -1,6 +1,0 @@
-import { DirResult, FileResult } from "tmp";
-
-/**
- * Provides information about a temporary filesystem-entry.
- */
-export type TempResult = FileResult | DirResult;