+++ src/tests/TempDirectory.test.ts
@@ -1,6 +1,6 @@
 import Assert = require("assert");
 import Path = require("path");
-import FileSystem = require("fs-extra");
+import { mkdir, pathExists, readFile, remove, stat, writeFile } from "fs-extra";
 import { TempDirectory } from "..";
 
 /**
@@ -13,7 +13,6 @@ export function TempDirectoryTests(): void
         () =>
         {
             let tempDir: TempDirectory;
-            let tempDirName: string;
             let tempFileName: string;
             let text: string;
 
@@ -24,37 +23,92 @@ export function TempDirectoryTests(): void
                     text = "Hello World";
                 });
 
+            setup(
+                () =>
+                {
+                    tempDir = new TempDirectory();
+                });
+
+            teardown(
+                () =>
+                {
+                    try
+                    {
+                        tempDir.Dispose();
+                    }
+                    catch { }
+                });
+
             suite(
                 "General",
                 () =>
                 {
                     test(
-                        "Checking whether a temporary directory can be created…",
-                        () =>
-                        {
-                            tempDir = new TempDirectory();
-                            tempDirName = tempDir.FullName;
-                        });
-
-                    test(
-                        "Checking whether the temporary directory exists…",
-                        async () => Assert.strictEqual(await FileSystem.pathExists(tempDirName), true));
+                        "Checking whether directories are created correctly…",
+                        async () => Assert.ok(await pathExists(tempDir.FullName)));
 
                     test(
                         "Checking whether files can be written inside the temporary directory…",
-                        async () => Assert.doesNotReject(async () => FileSystem.writeFile(tempDir.MakePath(tempFileName), text)));
+                        async () => Assert.doesNotReject(async () => writeFile(tempDir.MakePath(tempFileName), text)));
 
                     test(
-                        "Checking whether the file written inside the temporary directory exists…",
-                        async () => Assert.strictEqual((await FileSystem.readFile(tempDir.MakePath(tempFileName))).toString(), text));
+                        "Checking whether files written inside the temporary directory exists…",
+                        async () =>
+                        {
+                            await writeFile(tempDir.MakePath(tempFileName), text);
+                            Assert.strictEqual((await readFile(tempDir.MakePath(tempFileName))).toString(), text);
+                        });
 
                     test(
-                        "Checking whether the temporary directory can be disposed…",
+                        "Checking whether temporary directories can be disposed…",
                         () => Assert.doesNotThrow(() => tempDir.Dispose()));
 
                     test(
-                        "Checking whether the temporary directory has been deleted…",
-                        async () => Assert.strictEqual(await FileSystem.pathExists(tempDirName), false));
+                        "Checking whether temporary directories are deleted by invoking `Dispose`…",
+                        async () =>
+                        {
+                            tempDir.Dispose();
+                            Assert.ok(!await pathExists(tempDir.FullName));
+                        });
+
+                    test(
+                        "Checking whether temporary directories can be deleted even if they contain files…",
+                        async () =>
+                        {
+                            Assert.ok(await pathExists(tempDir.FullName));
+                            await writeFile(tempDir.MakePath(tempFileName), text);
+                            Assert.ok(await pathExists(tempDir.MakePath(tempFileName)));
+                            tempDir.Dispose();
+                            Assert.ok(!await pathExists(tempDir.MakePath(tempFileName)));
+                            Assert.ok(!await pathExists(tempDir.FullName));
+                        });
+
+                    test(
+                        "Checking whether the file-system entry name is available right after the disposal…",
+                        async () =>
+                        {
+                            tempDir.Dispose();
+                            await Assert.doesNotReject(() => mkdir(tempDir.FullName));
+                            return remove(tempDir.FullName);
+                        });
+
+                    if (process.platform === "linux")
+                    {
+                        test(
+                            "Checking whether the `mode` is applied correctly…",
+                            async () =>
+                            {
+                                let mode = 0o007;
+
+                                let dir = new TempDirectory(
+                                    {
+                                        Mode: mode
+                                    });
+
+                                Assert.strictEqual((await stat(dir.FullName)).mode & 0o777, mode);
+                                dir.Dispose();
+                            });
+                    }
                 });
 
             suite(
@@ -62,6 +116,7 @@ export function TempDirectoryTests(): void
                 () =>
                 {
                     let path: string[];
+                    let tempDir: TempDirectory;
 
                     suiteSetup(
                         () =>