+++ src/tests/createTempFile.ts
@@ -3,11 +3,12 @@ import { TempDirectory } from "../TempDirectory";
 import { TempFile } from "../TempFile";
 import { ITestFiles } from "./ITestFiles";
 
-let tempDir = new TempDirectory();
+let keep = process.argv.slice(1).some((argument) => argument === "keep");
+let tempDir = new TempDirectory({ Keep: keep });
 writeFileSync(tempDir.MakePath("Test"), "hello world");
 
 process.send(
     {
-        file: new TempFile().FullName,
+        file: new TempFile({ Keep: keep }).FullName,
         dir: tempDir.FullName
     } as ITestFiles);